feat: 🎸 added http client fo loin entity    #1
